Use material components
