+++ toggler/src/main/res/layout/layout_checkbox_toggle.xml
@@ -81,7 +81,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/lblValues"
                     tools:text ="Local Value"/>
-                <Switch
+                <com.google.android.material.switchmaterial.SwitchMaterial
                     android:id="@+id/swithcLocalProvider"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
